Android: Reset update progress when booted into new version.
In Android we don't reset progress after update completes because if
automatic update is toggled off after that, we will need to revert the
slot switch and still keep the progress.

However if we try to apply the same full update again after booted
into the new version, it will try to resume and fail on verification
because the inactive slot is now the previous version, so we shouldn't
resume in this case.

Bug: 122117961
Test: progress is reset after booted into new version
Change-Id: I394a78fd37bbbdcf62bf85639a9f86e7d3bc42c6
Signed-off-by: spezi77 <spezi7713@gmx.net>